move proofs related common tests
